fixed header + refs
