+++ src/js/support-ukraine.js
@@ -1,20 +1,18 @@
 import { funds } from './funds';
 
-// const list_support = document.querySelector('.js-support_list');
+const list_support = document.querySelector('.support-wrapper');
 
-// function createFondMarkup(arr) {
-//   const markup = arr
-//     .map(
-//       ({ number, img, url, title }) =>
-//         `<li>
-//     <a class="link support_funds_element" href="${url}">
-//     <span class="support_funds_number">${number}</span>
-//     <img class="support_funds_image" src="${img}" target="_blank" alt="${title}">
-// </a>
-// </li>`
-//     )
-//     .join('');
-//   return markup;
-// }
+function createFondMarkup(arr) {
+  const markup = arr
+    .map(
+      ({ number, img, url, title }) =>
+        `<div class="swiper-slide support-cont-slide"><a class="link support_funds_element" href="${url}" target="_blank" rel="noopener noreferrer">
+    <span class="support_funds_number">${number}</span>
+    <img class="support_funds_image" src="${img}" alt="${title}">
+</a></div>`
+    )
+    .join('');
+  return markup;
+}
 
-// list_support.insertAdjacentHTML('beforeend', createFondMarkup(funds));
+list_support.insertAdjacentHTML('beforeend', createFondMarkup(funds));
