+++ src/sass/components/_theme.scss
@@ -11,9 +11,21 @@ body.dark-theme {
 
 .dark-theme_header {
   background-color: $black-color;
-  color: $white-color;
 }
 
+.dark-theme_header-hover {
+  color: $white-color;
+  fill: $white-color;
+  &:hover,
+  &:focus {
+    color: $black-color;
+    fill: $black-color;
+  }
+}
+
+.dark-theme-cuttent-page {
+  color: $black-color;
+}
 .dark-theme-category-list_item {
   color: rgba(255, 255, 255, 0.6);
 }
